get id kontrol altında
autocomplete pass off

+++ delete.php
@@ -1,11 +1,16 @@
 <?php
     //require('./readDataConfig.php');
 	include "config.php";
-    $id = $_GET['id'];
-    
-        $sql = ("DELETE FROM kullanicilar WHERE id=".$id);
-        $result = mysqli_query($link, $sql);
-                    
+
+    $id = filter_var($_GET['id'],FILTER_SANITIZE_NUMBER_INT);
+    if(!filter_var($id, FILTER_VALIDATE_INT)){
     	header("location:newReadUsers.php");
+    	die();
+    } 
+
+    $sql = ("DELETE FROM kullanicilar WHERE id=".$id);
+    $result = mysqli_query($link, $sql);
+                    
+    header("location:newReadUsers.php");
  
 ?>
